Get the EC2 region list dynamically at run time
This replaces the compiled-in region list with one generated via the
DescribeRegions EC2 API call, in order to ensure that when AMIs are
published they are copied into all of the available EC2 regions and
not merely those which existed when the code was last updated.

Since this requires a new API call (DescribeRegions) it may be necesary
to update IAM rules; so the version number is being bumped to 1.1.0.