Retry some failed EC2 API calls
Specifically: DescribeRegions, DescribeConversionTasks,
DescribeSnapshots, DescribeImages, and ModifyImageAttributes.
These are all idempotent, so it is safe to retry them; and the
DescibeConversionTasks and DescribeSnapshots APIs are called
repeatedly (while waiting for a volume import and a snapshot
creation to complete respectively) so they're where the upload
is most likely to fail due to a networking or EC2 glitch.

If the ImportVolume, CreateSnapshot, DeleteVolumes, RegisterImage,
or CopyImage API calls fail, the upload will fail without retrying
them; but if any of those APIs fail we don't know what state we're
in so it's safest to fail and let a human clean up the mess.