Ignore SIGPIPE globally.
OpenSSL doesn't know how to use MSG_NOSIGNAL when writing to sockets;
as a workaround, we were setting SO_NOSIGPIPE on sockets before
handing them off to OpenSSL.  Unfortunately SO_NOSIGPIPE is not
portable, so this failed on non-BSD platforms.

Fortunately the only place where we could encounter SIGPIPE is from
a network socket -- we're not generating large amounts of output to
stdout -- so ignoring SIGPIPE globally is adequate, albeit somewhat
ham-fisted.