Work around ImportVolume request limits
Add a new ec2_apicall_resourcelimited function which is like ec2_apicall
except that if a request fails with ResourceLimitExceeded, it waits a
minute and then tries again.  Use it in importvolume, because FreeBSD
now tries to publish all of the AMIs at once (8 on 15.x, 8 on 14.x, and
2 on 13.x) and most of the uploads failed.

While I'm here, fix some whitespace bugs.