Detect SERVER_ERROR when importing volume
Around 1% of EBS volume imports are failing with a status message of
"SERVER_ERROR : conversion : AWS initiated task cancellation"; prior
to this change, the upload hangs (without exiting) because we're
waiting for a volume ID to be provided.  As a practical matter, this
results in the FreeBSD release build process hanging since it waits
for all of the uploads to "complete" (aka for all the processes to
exit).

Check for the substring "SERVER_ERROR" in the status message and
error out if encountered.

Sponsored by:	Amazon